chore : make general changes, add more interfaces and aplied in code

+++ src/use-cases/user/deleteUserUseCase.ts
@@ -1,15 +1,25 @@
 import { PrismaUserRepository } from '../../infra/adapters/repositories/prisma/prisma-user-repository';
-import { IUser } from '../../utils/@interfaces';
+import { UserResponse } from '../../utils/@interfaces';
+import { MissingParamError, ParamDoesNotExist } from '../../utils/errors/index.';
+import { GetUserByIdUseCase } from './getUserUseCase';
 
 export class DeleteUserUseCase {
-  constructor(private deleteUser: PrismaUserRepository) {}
+  constructor(
+    private deleteUser: PrismaUserRepository,
+    private userService : GetUserByIdUseCase
+  ) {}
 
-  async execute(id: string) : Promise<IUser | null> {
+  async execute(id: string): Promise<UserResponse | void> {
+    const user = await this.userService.findUserById(id);
+
     if (!id) {
-      throw new Error('id not found');
+      throw new MissingParamError('user id not found on request');
     }
-    const deleteUser = await this.deleteUser.findByIdAndDelete(id);
 
-    return deleteUser;
+    if (!user) {
+      throw new ParamDoesNotExist(`this user id ${id} does not exist`);
+    }
+
+    await this.deleteUser.findByIdAndDelete(id);
   }
 }
